moved optional_auth to auth.py (does not work at the moment)
